requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
fastapi==0.109.0
Flask==3.0.0
Flask>=3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.2
openai==1.12.0
openai>=2.0.0
pandas==3.0.0
pandas>=2.0.0
pygithub==2.1.1
PyGithub
pydantic==2.5.0
PyPDF2==3.0.1
PyPDF2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-dotenv>=1.0.0
python-multipart==0.0.6
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
